Simple model ft_printf

+++ ft_printf/turn_in/libft/ft_putchar_fd.c
@@ -6,13 +6,13 @@
 /*   By: ajacome- <ajacome-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/14 12:35:39 by andresj           #+#    #+#             */
-/*   Updated: 2023/07/17 15:56:18 by ajacome-         ###   ########.fr       */
+/*   Updated: 2023/07/18 10:13:16 by ajacome-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_putchar_fd(char c, int fd)
+ssize_t	ft_putchar_fd(char c, int fd)
 {
 	return (write(fd, &c, 1));
 }
